yvterm + ygui2: the drawable contract end to end — rolling rich store, retained widget toolkit, multi-language FFI (#732)
Terminal-side engine (yvterm):
- Rolling rich store: addressable rich content on the terminal wire —
  blocks, nested groups and own-id complexes, addressed by a 64-bit
  folded id path; insert (exact-subtree replacement), update, delete,
  CMD_PATH at any depth; wire-order, command-local application.
- Group offsets are the ONE mover (a ~20-byte update scrolls an app's
  whole surface); declared row spans + CMD_RESERVE; per-NODE scroll
  retirement with sealing per block; journaled complex updates replayed
  on re-materialization; group-slot and record reclamation keep
  long-lived insertions bounded (one reachability scan per replacement
  window). The normative model lives in drawable-use-cases.md.
- Unified render: ONE total paint order over SDF prims, shaped text and
  own-pipeline complexes (paint_z, sequence, ordinal), with a C<->WGSL
  layout contract test.

One committed scale (HiDPI + structural zoom):
- The wire is PRODUCER-LOGICAL pixels; the receiver converts by one
  committed product (display density x structural cell zoom) at every
  consumer: reserve/span conversion, retirement, SDF bucketing and
  sampling (framebuffer row anchors, scaled local space; shaped
  terminal glyphs are a scale-1 base range), complex placement and
  group offsets. Live density transitions update renderer, grid and
  baseline atomically; footprint row counts are invariant under both
  density and zoom.
- The pane resize contract publishes the exact committed snapped
  product; a structural zoom tick reflows an attached client like
  terminal text, with one PTY resize per transition; ygui2 accepts the
  host's full positive-finite scale range transactionally.

Non-destructive complex resize:
- A live figure resize is ONE ~28-byte addressed geometry op: the
  retained runtime mutates in place, the record re-plans locally
  (yplot rechrome: bounds/ticks in the uniform words, chrome tail),
  the receiver replaces ONLY the figure's chrome group at its original
  paint z, and the retirement extent follows the new size. The record
  and its accumulated data are never re-shipped; failed chrome
  replacements retry at every envelope close.

ygui2 — retained widget toolkit on the drawable contract:
- A pure PTY client (DCS drawables out, client-input OSC in): widget
  tree = wire group tree; five dirty classes emit the smallest
  sufficient op (offset move, skin reopen, subtree reopen, addressed
  complex stream); clean frames ship zero bytes.
- Full widget set incl. scrollarea (offset-only scrolling with a
  measured clamp), overlays/dialogs/popups, focus traversal, plot
  widget with streamed ring buffers (~40 bytes/sample), complex host
  and ydraw embed; fullscreen and inline reservation modes; the
  yguiapp2 host loop (raw termios, HOLD/ACK exit barrier); ytop2 and
  ygreeter2 ports.

FFI, four languages:
- Ergonomic ygui2 wrappers for Python (ctypes), Lua (LuaJIT ffi), Go
  (cgo) and TypeScript (koffi) with the same shape: App host loop, Node
  builders, callback trampolines, RadioGroup, deferred callback-time
  close, Result cause-chain ownership, reservation-mode + content-scale
  surface. Demos hello/counter/catalog/dashboard in every language plus
  pane launcher scripts; generated v2 drawable bindings refreshed.

Tests:
- yvterm: rich lifecycle (retirement boundaries, density/zoom
  transitions over retained blocks, recycling/topology boundedness,
  transform contract), scroll tiers, wire semantics.
- ygui2: wire-cost contract; Python/Lua/Go/TypeScript binding suites
  (callback-time close, error ownership, mode parity, explicit-zero).
- yterminal: headless ingest harness (#728) — the producer→terminal
  roundtrip (runtime identity, extent refresh, chrome replacement and
  retry, HiDPI scales, retained-block density transition) and the
  zoom→envelope→client resize contract on the production view-event
  path.
- yplot: emit_into/emit_expression parity, rechrome and atomic ranges.

+++ src/yetty/ydrawlist2/drawable.c
@@ -146,7 +146,9 @@ struct YETTY_ANNOTATE("class@ydrawlist2:text") YETTY_ANNOTATE("parent@ydrawlist2
     YETTY_ANNOTATE("property") float y;
     YETTY_ANNOTATE("property") float font_size;
     YETTY_ANNOTATE("property") uint32_t color;
-    YETTY_ANNOTATE("property") uint32_t layer;
+    /* Stacking depth (z-order), uniform and SIGNED across every drawable's
+     * `layer` — negative layers paint below the default plane. */
+    YETTY_ANNOTATE("property") int32_t layer;
     YETTY_ANNOTATE("property") int32_t font_id;
     YETTY_ANNOTATE("property") float rotation;
     uint32_t body_len;
@@ -225,9 +227,9 @@ static struct yetty_ycore_void_result text_pack(struct yetty_yclass_object *obj,
         .size = text->body_len,
         .capacity = text->body_len,
     };
-    struct yetty_ycore_void_result pack_r =
-        yetty_ydraw_drawable_list_add_text(list, text->x, text->y, &body, text->font_size,
-                                           text->color, text->layer, text->font_id, text->rotation);
+    struct yetty_ycore_void_result pack_r = yetty_ydraw_drawable_list_add_text(
+        list, text->x, text->y, &body, text->font_size, text->color, (uint32_t)text->layer,
+        text->font_id, text->rotation);
     YETTY_RETURN_IF_ERR(yetty_ycore_void, pack_r, "ydrawlist2 text pack: add_text");
     return YETTY_OK_VOID();
 }
